lint only user-supplied files
