+++ src/main/kotlin/com/parmet/buf/gradle/BufExtension.kt
@@ -44,7 +44,7 @@ open class BufExtension {
     /**
      * Specify the version of Buf.
      */
-    var toolVersion: String = "1.1.0"
+    var toolVersion: String = "1.3.1"
 
     internal var imageArtifactDetails: ArtifactDetails? = null
