+++ src/main/kotlin/com/parmet/buf/gradle/DockerSupport.kt
@@ -28,17 +28,14 @@ internal val BUF_DOCKER_TASK_DEPENDENCIES =
 
 internal fun Exec.buf(ext: BufExtension, vararg args: Any) {
     dependsOn(BUF_DOCKER_TASK_DEPENDENCIES)
+    buf(project, ext, args.asList())
+}
 
+internal fun ExecSpec.buf(project: Project, ext: BufExtension, args: Iterable<Any>) {
     commandLine("docker")
     val dockerArgs = project.baseDockerArgs(ext) + args
     setArgs(dockerArgs)
-    project.logger.quiet("Running buf: `docker ${dockerArgs.joinToString(" ")}`")
-}
-
-internal fun ExecSpec.bufLint(project: Project, ext: BufExtension, vararg args: Any) {
-    commandLine("docker")
-    setArgs(project.lintDockerArgs(ext) + args)
-    project.logger.quiet("Running buf: `docker ${getArgs().joinToString(" ")}`")
+    project.logger.debug("Running buf: `docker ${dockerArgs.joinToString(" ")}`")
 }
 
 private fun Project.baseDockerArgs(ext: BufExtension) =
@@ -47,14 +44,5 @@ private fun Project.baseDockerArgs(ext: BufExtension) =
         "--rm",
         "--volume", "$projectDir:/workspace:Z",
         "--workdir", "/workspace/build/bufbuild",
-        "bufbuild/buf:${ext.toolVersion}"
-    )
-
-private fun Project.lintDockerArgs(ext: BufExtension) =
-    listOf(
-        "run",
-        "--rm",
-        "--volume", "$projectDir:/workspace:Z",
-        "--workdir", "/workspace",
         "bufbuild/buf:${ext.toolVersion}"
     )
