+++ src/test/resources/LintTest/linting with a config in default location/buf.yaml
@@ -13,8 +13,6 @@
 
 version: v1
 lint:
-  ignore:
-    - google
   use:
     - DEFAULT
   except:
